+++ apps/ethereum/tribes/src/pages/_app.tsx
@@ -1,17 +1,19 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-import { initialize, Provider} from "@decentology/hyperverse";
+import { initialize, Provider } from "@decentology/hyperverse";
 import { networks } from "@decentology/hyperverse";
 import { Ethereum } from "@decentology/hyperverse-ethereum";
 import * as Tribes from "@decentology/hyperverse-ethereum-tribes";
 import InnerComponent from "../components/InnerComponent";
-import { ToastContainer } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
-
+import { ToastContainer } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 function MyApp({ Component, pageProps }: AppProps) {
   const hyperverse = initialize({
     blockchain: Ethereum,
     network: networks.Testnet,
+    storage: {
+      options: { clientUrl: "https://fileportal.org" }, // Updated to fileportal due to SSL error issues with SiaSky.net
+    },
     modules: [
       {
         bundle: Tribes,