+++ apps/ethereum/tribes/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @decentology/samples-web-tribes
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @decentology/hyperverse@1.0.7
+  - @decentology/hyperverse-ethereum@1.0.7
+  - @decentology/hyperverse-ethereum-tribes@1.0.7
+
 ## 1.0.4
 
 ### Patch Changes